Fix JSON schema patterns for surrogates
JSON operates on UTF-16 strings. No UTF-32 is supported yet. Hence, we
need to fix the patterns to operate on UTF-16 code points, instead of
using UTF-32 code points.
